#38
- preset/bun: добавлен sentry
- preset/bun: утилиты вынесены из провайдеров
- preset/bun: логирование исходящих запросов теперь не учитывает StatusError
