+++ src/preset/bun/handler/utils/index.ts
@@ -1,7 +1,9 @@
 /* eslint-disable require-jsdoc, jsdoc/require-jsdoc */
+import type { Handler } from '../../../../http';
+import type { BaseConfig } from '../../../../config';
 import { KnownToken } from '../../../../tokens';
 import { CURRENT_APP, type Application, type Resolve } from '../../../../di';
-import type { Handler } from '../../../../http';
+import { getClientIp } from '../../bun/utils';
 
 export function HandlerProvider(getApp: () => Application) {
   return (resolve: Resolve): Handler => {
@@ -16,4 +18,40 @@ export function HandlerProvider(getApp: () => Application) {
       return app.get(KnownToken.Http.Handler.main)(request);
     };
   };
+}
+
+export function getResponseFormat(request: Request): 'html' | 'json' {
+  let result: 'html' | 'json' = 'html';
+
+  if ((request.headers.get('accept') || '').toLowerCase().includes('application/json')) {
+    result = 'json';
+  }
+
+  return result;
+}
+
+export function getForwardedHeaders(config: BaseConfig, request: Request): Headers {
+  const result = new Headers();
+
+  // user agent
+  result.set('User-Agent', `simaland-${config.appName}/${config.appVersion}`);
+
+  // client ip
+  const clientIp = getClientIp(request);
+
+  if (clientIp) {
+    result.set('X-Client-Ip', clientIp);
+  }
+
+  // service headers
+  request.headers.forEach((headerValue, headerName) => {
+    if (
+      headerName.toLowerCase().startsWith('simaland-') &&
+      headerName.toLowerCase() !== 'simaland-params'
+    ) {
+      result.set(headerName, headerValue);
+    }
+  });
+
+  return result;
 }
